+++ src/singlewinner/cardinal/asm.cc
@@ -15,7 +15,7 @@ std::pair<ordering, bool> app_sorted_margins::elect_inner(
 	condmat pairwise_matrix(papers, num_candidates, CM_PAIRWISE_OPP);
 
 	std::vector<double> approvals(num_candidates, 0);
-	std::vector<std::pair<double, int> > approvals_by_cand;
+	std::vector<std::pair<double, size_t> > approvals_by_cand;
 
 	// Use cardinal ratings to get the scores.
 	ordering approval_outcome = approval_filter->elect(
@@ -49,8 +49,8 @@ std::pair<ordering, bool> app_sorted_margins::elect_inner(
 
 		// Look for adjacent candidates that are out of order.
 		for (size_t i = 0; i < num_hopefuls-1; ++i) {
-			int incumbent = approvals_by_cand[i].second,
-				challenger = approvals_by_cand[i+1].second;
+			size_t incumbent = approvals_by_cand[i].second,
+				   challenger = approvals_by_cand[i+1].second;
 
 			if (!pairwise_matrix.beats(challenger, incumbent)) {
 				continue;
